+++ sentry-android-core/src/main/java/io/sentry/android/core/ContextUtils.java
@@ -351,7 +351,7 @@ public final class ContextUtils {
     }
   }
 
-  /** Register a not exported BroadcastReceiver, independently from platform version. */
+  /** Register an exported BroadcastReceiver, independently from platform version. */
   static @Nullable Intent registerReceiver(
       final @NotNull Context context,
       final @NotNull SentryOptions options,
@@ -360,7 +360,7 @@ public final class ContextUtils {
     return registerReceiver(context, new BuildInfoProvider(options.getLogger()), receiver, filter);
   }
 
-  /** Register a not exported BroadcastReceiver, independently from platform version. */
+  /** Register an exported BroadcastReceiver, independently from platform version. */
   @SuppressLint({"NewApi", "UnspecifiedRegisterReceiverFlag"})
   static @Nullable Intent registerReceiver(
       final @NotNull Context context,